Install statgarten dependencies.

Install statgarten dependencies, datatoys, board, scissor, colorpen, soroban, stove

Usage

init()

Value

an HTML list

Examples

init()
#> Downloading GitHub repo statgarten/datatoys@HEAD
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/private/var/folders/xs/53yprhh96l72pgw833r4thv40000gn/T/Rtmp8LhMfm/remotes14cfc40eb19f/statgarten-datatoys-d6226a7/DESCRIPTION’ ... OK
#> * preparing ‘datatoys’:
#> * checking DESCRIPTION meta-information ... OK
#> * excluding invalid files
#> Subdirectory 'R' contains invalid file names:
#>   ‘.Rapp 2.history’
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> * building ‘datatoys_0.1.0.tar.gz’
#> 
#> Installing package into ‘/private/var/folders/xs/53yprhh96l72pgw833r4thv40000gn/T/RtmpCekCyl/temp_libpath1417f69dcd2ce’
#> (as ‘lib’ is unspecified)
#> Downloading GitHub repo statgarten/board@HEAD
#> 
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/private/var/folders/xs/53yprhh96l72pgw833r4thv40000gn/T/Rtmp8LhMfm/remotes14cfc68af7aad/statgarten-board-594c9bc/DESCRIPTION’ ... OK
#> * preparing ‘board’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> Removed empty directory ‘board/vignettes’
#> * building ‘board_0.0.4.tar.gz’
#> 
#> Installing package into ‘/private/var/folders/xs/53yprhh96l72pgw833r4thv40000gn/T/RtmpCekCyl/temp_libpath1417f69dcd2ce’
#> (as ‘lib’ is unspecified)
#> Downloading GitHub repo statgarten/scissor@HEAD
#> 
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/private/var/folders/xs/53yprhh96l72pgw833r4thv40000gn/T/Rtmp8LhMfm/remotes14cfc290b36a4/statgarten-scissor-49bb9df/DESCRIPTION’ ... OK
#> * preparing ‘scissor’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> Removed empty directory ‘scissor/vignettes’
#> * building ‘scissor_0.2.3.tar.gz’
#> 
#> Installing package into ‘/private/var/folders/xs/53yprhh96l72pgw833r4thv40000gn/T/RtmpCekCyl/temp_libpath1417f69dcd2ce’
#> (as ‘lib’ is unspecified)
#> Downloading GitHub repo statgarten/colorpen@HEAD
#> Error: Failed to install 'colorpen' from GitHub:
#>   HTTP error 403.
#>   API rate limit exceeded for 221.151.83.21. (But here's the good news: Authenticated requests get a higher rate limit. Check out the documentation for more details.)
#> 
#>   Rate limit remaining: 0/60
#>   Rate limit reset at: 2023-10-29 04:15:13 UTC
#> 
#>   To increase your GitHub API rate limit
#>   - Use `usethis::create_github_token()` to create a Personal Access Token.
#>   - Use `usethis::edit_r_environ()` and add the token as `GITHUB_PAT`.